+++ Wrap/python/__init__.py.in
@@ -11,3 +11,5 @@ from libBornAgainCore import *
 sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 
 from plot_utils import *
+from yaml_utils import *
+
